+++ History.md
@@ -1,9 +1,11 @@
-## v0.1.2 - unreleased
+## v0.1.2 - April 4, 2014
 
  * Allow `.push(name)` without a function to switch to previously-defined panes.  
  This makes `.go()` deprecated.
 
  * Purge obsolete panes.
+
+ * Prevent scrolling during transitions.
 
 ## v0.1.1 - March 26, 2014